Teach snarl about --overwrite and avoid overwriting files
Snarl will no longer overwrite an existing file during a tangle run
unless you specify --overwrite (-w). This prevents accidental loss of
data due to a conflict between a :file block in your source and a file
on disk.

This also introduces the 'snarl files' command, which will list the
files contained in a source document.

Closes #1